% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{gkg}
\alias{gkg}
\title{Google Knowledge Graph}
\usage{
gkg(q, limit = 1L, languages = NULL, types = NULL, prefix = NULL,
  key = Sys.getenv("GKG_API_KEY"))
}
\arguments{
\item{q}{The search query.}

\item{limit}{The maximum number of results to retrieve.}

\item{languages}{The list of language codes (defined in
ISO 639) to run the query with, for instance \code{en}.}

\item{types}{Restricts returned entities to those of the 
specified types. For example, you can specify \code{Person}
(as defined in \url{http://schema.org/Person}) to restrict
the results to entities representing people. If multiple 
types are specified, returned entities will contain one or 
more of these types.}

\item{prefix}{Enables prefix (initial substring) match 
against names and aliases of entities. For example, a prefix
\code{Jung} will match entities and aliases such as \code{Jung},
\code{Jungle}, and \code{Jung-ho Kang}.}

\item{key}{Your API key, defaults to looking for the 
\code{GKG_API_KEY} environment variable.}
}
\description{
Call the Google Knowledge Graph API.
}
